Feature: TableParitionWriter support parition reloption
Implement partition writer in table and micro-partition layer, which
will be able to be divided into multiple files when writing in PAX,
which will make min/max statistics more effective.

  - Support MicroPartitionWriter and MicroPartitionStats merge
  - Support partition writer which used to support partition_by options

+++ contrib/pax_storage/src/cpp/storage/columns/pax_encoding_column.cc
@@ -6,7 +6,7 @@ namespace pax {
 
 template <typename T>
 PaxEncodingColumn<T>::PaxEncodingColumn(
-    uint64 capacity, const PaxEncoder::EncodingOption &encoding_option)
+    uint32 capacity, const PaxEncoder::EncodingOption &encoding_option)
     : PaxCommColumn<T>(capacity),
       encoder_options_(encoding_option),
       encoder_(nullptr),
@@ -19,7 +19,7 @@ PaxEncodingColumn<T>::PaxEncodingColumn(
 
 template <typename T>
 PaxEncodingColumn<T>::PaxEncodingColumn(
-    uint64 capacity, const PaxDecoder::DecodingOption &decoding_option)
+    uint32 capacity, const PaxDecoder::DecodingOption &decoding_option)
     : PaxCommColumn<T>(capacity),
       encoder_(nullptr),
       decoder_options_{decoding_option},
